backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.0.0
sentence-transformers>=2.7.0,<3
jira>=3.6.0,<4
requests>=2.28.0,<3
atlassian-python-api>=3.41.0,<4
beautifulsoup4>=4.12.3,<5
langchain-core==0.3.76
langchain-community==0.3.27
langchain-text-splitters==0.3.9
langchain-aws>=0.2.30
boto3>=1.39.7
pypdf>=4.2.0,<5
pandas>=2.2.0,<3
numpy>=1.26,<2
openpyxl>=3.1,<4
chromadb>=0.5.5,<0.6
python-docx>=1.1.0,<1.2
markdown>=3.5,<4
pytest>=7.4,<9
